Moved the validation of input and output locations from the annotation handlers to the property spec implementations. This change means that the annotation handlers do not have to duplicate the logic for converting the property value to a file or set of files in order to validate the value.
It also means that the validation could later be applied to inputs and output defined using the APIs rather than define via annotated properties. It is not yet applied to these.

This change introduces some duplication of the validation logic in the various property spec implementations. This will be cleaned up later.